feat: add must getters
Add getters which panics if key is not found or value returned is default data type.
For example: `kf.MustInt("abc")`
If path "abc" is not found or its 0 then it panics else returnes value.
